管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -42,6 +42,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかチェック
+        if(checkAdmin()) {
+
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -63,7 +66,7 @@ public class EmployeeAction extends ActionBase {
 
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
-
+        }
     }
 
     /**
@@ -73,11 +76,14 @@ public class EmployeeAction extends ActionBase {
      */
     public void entryNew() throws ServletException, IOException{
 
+      //管理者かどうかチェック
+        if(checkAdmin()) {
         putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
         //新規登録画面表示
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /**
@@ -88,7 +94,7 @@ public class EmployeeAction extends ActionBase {
     public void create() throws ServletException, IOException{
 
         //CSRF対策 tokenのチェック
-        if(checkToken()) {
+        if(checkToken() && checkAdmin()) {
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -135,6 +141,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void show() throws ServletException, IOException {
 
+        //管理者かどうかチェック
+        if(checkAdmin()) {
+
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -149,6 +158,7 @@ public class EmployeeAction extends ActionBase {
 
         //詳細画面を表示
         forward(ForwardConst.FW_EMP_SHOW);
+        }
     }
 
     /**
@@ -157,6 +167,9 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void edit() throws ServletException, IOException {
+
+        //管理者かどうかチェック
+        if(checkAdmin()) {
 
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -173,6 +186,7 @@ public class EmployeeAction extends ActionBase {
 
     //編集画面を表示する
     forward(ForwardConst.FW_EMP_EDIT);
+        }
     }
 
     /**
@@ -182,8 +196,8 @@ public class EmployeeAction extends ActionBase {
      */
     public void update() throws ServletException, IOException{
 
-        //SCRF対策 tokenのチェック
-        if(checkToken()) {
+        //管理者 SCRF対策 tokenのチェック
+        if(checkAdmin() && checkToken()) {
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -229,8 +243,8 @@ public class EmployeeAction extends ActionBase {
      */
     public void destroy() throws ServletException, IOException {
 
-      //CSRF対策 tokenのチェック
-        if (checkToken()) {
+      //管理者 CSRF対策 tokenのチェック
+        if (checkAdmin() && checkToken()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -240,6 +254,28 @@ public class EmployeeAction extends ActionBase {
 
             //一覧画面にリダイレクト
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    public boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+            return true;
         }
     }
